+++ build_runner_core/lib/src/generate/build_result.dart
@@ -54,10 +54,12 @@ enum BuildStatus {
 
 /// The type of failure
 class FailureType {
-  static const general = FailureType._(1);
-  static const cantCreate = FailureType._(73);
+  static final general = FailureType._(1);
+  static final cantCreate = FailureType._(73);
+  static final buildConfigChanged = FailureType._(75);
+  static final buildScriptChanged = FailureType._(75);
   final int exitCode;
-  const FailureType._(this.exitCode);
+  FailureType._(this.exitCode);
 }
 
 abstract class BuildState {